improved component structuring
