+++ src/tool/index.tsx
@@ -1,0 +1,1 @@
+export { default as Weights } from './weights';
