+++ src/views/Home.tsx
@@ -2,7 +2,7 @@ import { defineComponent } from 'vue';
 import * as store from '../store';
 import assert from '../util/assert';
 import Row from '../components/row';
-import { BMI } from '../util/calculators';
+import * as tools from '../tool';
 
 export default defineComponent({
   name: 'Home',
@@ -14,7 +14,7 @@ export default defineComponent({
     return (
       <div class="home">
         Home component {weight.s} {height.s}
-        <Row name="Example">{BMI(weight, height).s}</Row>
+        <tools.Weights></tools.Weights>
       </div>
     );
   }
